removed the tomcat embedded dependencies so the project will run on the reactive netty embedded server

+++ src/main/java/com/compomics/secretesite/controllers/repositories/SpeciesRepository.java
@@ -8,5 +8,5 @@ import org.springframework.data.rest.core.annotation.RepositoryRestResource;
  * Created by davy on 4/11/2017.
  */
 @RepositoryRestResource
-public interface SpeciesRepository extends PagingAndSortingRepository<Species, Long> {
+public interface SpeciesRepository extends PagingAndSortingRepository<Species, Integer> {
 }
